move MerchantIntegration enum to separate file
